% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshPlot.R
\name{cshPlot}
\alias{cshPlot}
\title{Produces plot of a sequence's elements and value}
\usage{
cshPlot(df)
}
\arguments{
\item{df}{a data frame consisting of four columns: first three as elements in a sequence and the
fourth as the final element n.}
}
\value{
a line plot of each element in a recursive sequence and its respective value
}
\description{
This function takes a data frame of four columns. The first three act as columns act as the first three
numbers of a sequence and the final column is the target nth element of the sequence. The function will then
find the value of all n elements in the sequence, and produce a line plot of its value against element n.
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)

cshPlot(my_data)
}
