% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshSequence.R
\name{cshSequence}
\alias{cshSequence}
\title{Finds the nth element of a recursive sequence}
\usage{
cshSequence(x, n)
}
\arguments{
\item{x}{a vector of three numbers that acts as the beginning of a sequence}

\item{n}{an integer denoting the nth element of a sequence}
}
\value{
the value of the nth element of the recursive sequence
}
\description{
The function takes two parameters: a vector of three numbers and an integer denoting the nth element of a recursive
sequence. The function will find the value of the nth element in a specific recursive sequence.
}
\examples{
cshSequence(c(2, 4, 3), 5)
}
