% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pcshoffmann-package.R
\docType{package}
\name{hw04pcshoffmann-package}
\alias{hw04pcshoffmann}
\alias{hw04pcshoffmann-package}
\title{hw04pcshoffmann: What the Package Does (One Line, Title Case)}
\description{
Package intended for hw04 of STAT-413 (Data Science).
}
\author{
\strong{Maintainer}: Christopher Hoffmann \email{ch7710a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
